1_sla_monitoring/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
apscheduler
watchdog
httpx
streamlit
pyyaml